+++ stories/components/data-table/tests/examples/accordion-tests/accrodion-with-more-interactive.tsx
@@ -51,9 +51,8 @@ const Demo = () => {
                 title='Go to our awesome article'
                 data-test-id='interactive-icon'
                 color='icon-secondary-neutral'
-                onClick={stopPropagation}
               />
-              <DescriptionTooltip onClick={stopPropagation}>
+              <DescriptionTooltip>
                 <DescriptionTooltip.Trigger tag={ButtonLink} use='secondary'>
                   About fastest animals
                 </DescriptionTooltip.Trigger>